# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for QuokkaQ, rewritten in Go.",
        "title": "QuokkaQ Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/units/{unitId}/shift/counters": {
            "get": {
                "description": "Retrieves counters for shift view",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Get shift counters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Counter"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/shift/eod": {
            "post": {
                "description": "Performs end of day operations for a unit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Execute End of Day",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/units/{unitId}/shift/queue": {
            "get": {
                "description": "Retrieves current queue tickets for a unit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shift"
                ],
                "summary": "Get queue tickets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit ID",
                        "name": "unitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Booking": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                },
                "userPhone": {
                    "type": "string"
                }
            }
        },
        "models.Counter": {
            "type": "object",
            "properties": {
                "assignedTo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
        "models.Service": {
            "type": "object",
            "properties": {
                "backgroundColor": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "description": {
                    "type": "string"
                },
                "descriptionEn": {
                    "type": "string"
                },
                "descriptionRu": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "gridCol": {
                    "type": "integer"
                },
                "gridColSpan": {
                    "type": "integer"
                },
                "gridRow": {
                    "description": "Grid configuration",
                    "type": "integer"
                },
                "gridRowSpan": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isLeaf": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                },
                "nameRu": {
                    "type": "string"
                },
                "numberSequence": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "prebook": {
                    "type": "boolean"
                },
                "prefix": {
                    "type": "string"
                },
                "textColor": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
        "models.Ticket": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/models.Booking"
                },
                "bookingId": {
                    "type": "string"
                },
                "calledAt": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "confirmedAt": {
                    "type": "string"
                },
                "counter": {
                    "$ref": "#/definitions/models.Counter"
                },
                "counterId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastCalledAt": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "queueNumber": {
                    "type": "string"
                },
                "service": {
                    "$ref": "#/definitions/models.Service"
                },
                "serviceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}